+++ lua/mini/align.lua
@@ -434,6 +434,15 @@ local H = {}
 ---   require('mini.align').setup({}) -- replace {} with your config table
 --- <
 MiniAlign.setup = function(config)
+  -- TODO: Remove after Neovim=0.9 support is dropped
+  if vim.fn.has('nvim-0.10') == 0 then
+    vim.notify(
+      '(mini.align) Neovim<0.10 is soft deprecated (module works but is not supported).'
+        .. " It will be deprecated after the next 'mini.nvim' release (module might not work)."
+        .. ' Please update your Neovim version.'
+    )
+  end
+
   -- Export module
   _G.MiniAlign = MiniAlign