+++ lua/mini/extra.lua
@@ -82,6 +82,15 @@ local H = {}
 ---   require('mini.extra').setup({}) -- replace {} with your config table
 --- <
 MiniExtra.setup = function(config)
+  -- TODO: Remove after Neovim=0.9 support is dropped
+  if vim.fn.has('nvim-0.10') == 0 then
+    vim.notify(
+      '(mini.extra) Neovim<0.10 is soft deprecated (module works but is not supported).'
+        .. " It will be deprecated after the next 'mini.nvim' release (module might not work)."
+        .. ' Please update your Neovim version.'
+    )
+  end
+
   -- Export module
   _G.MiniExtra = MiniExtra