+++ lua/mini/map.lua
@@ -193,6 +193,15 @@ local H = {}
 ---   require('mini.map').setup({}) -- replace {} with your config table
 --- <
 MiniMap.setup = function(config)
+  -- TODO: Remove after Neovim=0.9 support is dropped
+  if vim.fn.has('nvim-0.10') == 0 then
+    vim.notify(
+      '(mini.map) Neovim<0.10 is soft deprecated (module works but is not supported).'
+        .. " It will be deprecated after the next 'mini.nvim' release (module might not work)."
+        .. ' Please update your Neovim version.'
+    )
+  end
+
   -- Export module
   _G.MiniMap = MiniMap